Expand models on their own files.
